# Cars-api/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Car API",
      "version": "1.0.0",
      "description": "API for managing car records"
    },
    "servers": [
      {
        "url": "se-341-projects.onrender.com",
        "description": "Render Deployment"
      }
    ],
    "paths": {
      "/api/cars": {
        "get": {
          "summary": "Retrieve all cars",
          "responses": {
            "200": {
              "description": "A list of cars",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Car"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new car",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "responses": {
            "201": { "description": "Car added successfully" },
            "400": { "description": "Bad request" }
          }
        }
      },
      "/api/cars/{id}": {
        "put": {
          "summary": "Update a car",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Car updated successfully" },
            "404": { "description": "Car not found" }
          }
        },
        "delete": {
          "summary": "Delete a car",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": { "description": "Car deleted successfully" },
            "404": { "description": "Car not found" }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Car": {
          "type": "object",
          "properties": {
            "make": { "type": "string" },
            "model": { "type": "string" },
            "year": { "type": "integer" },
            "mileage": { "type": "integer" },
            "price": { "type": "integer" },
            "color": { "type": "string" }
          },
          "required": ["make", "model", "year", "mileage", "price", "color"]
        }
      }
    }
  }